Rename test file and class
